[docs] Move display title helper
